Address minor comments from review
